Adding MicrosoftAccount type superclass for MicrosoftStsAccount and AzureActiveDirectoryAccount
