+++ common/src/main/java/com/microsoft/identity/common/internal/providers/microsoft/microsoftsts/MicrosoftStsAccount.java
@@ -1,27 +1,16 @@
 package com.microsoft.identity.common.internal.providers.microsoft.microsoftsts;
 
 import com.microsoft.identity.common.adal.internal.util.StringExtensions;
-import com.microsoft.identity.common.internal.providers.microsoft.azureactivedirectory.AzureActiveDirectoryAccount;
+import com.microsoft.identity.common.internal.providers.microsoft.MicrosoftAccount;
 import com.microsoft.identity.common.internal.providers.microsoft.azureactivedirectory.ClientInfo;
 import com.microsoft.identity.common.internal.providers.oauth2.IDToken;
 
 import java.util.Map;
 
-public class MicrosoftStsAccount extends AzureActiveDirectoryAccount {
+public class MicrosoftStsAccount extends MicrosoftAccount {
 
     public MicrosoftStsAccount(IDToken idToken, String uid, final String uTid) {
         super(idToken, uid, uTid);
-    }
-
-    @Override
-    protected String getDisplayableId(Map<String, String> claims) {
-        if (!StringExtensions.isNullOrBlank(claims.get(MicrosoftStsIdToken.PREFERRED_USERNAME))) {
-            return claims.get(MicrosoftStsIdToken.PREFERRED_USERNAME);
-        } else if (!StringExtensions.isNullOrBlank(claims.get(MicrosoftStsIdToken.EMAIL))) {
-            return claims.get(MicrosoftStsIdToken.EMAIL);
-        }
-
-        return null;
     }
 
     /**
@@ -52,6 +41,17 @@ public class MicrosoftStsAccount extends AzureActiveDirectoryAccount {
     @Override
     public String getAuthorityType() {
         return "MSSTS";
+    }
+
+    @Override
+    protected String getDisplayableId(Map<String, String> claims) {
+        if (!StringExtensions.isNullOrBlank(claims.get(MicrosoftStsIdToken.PREFERRED_USERNAME))) {
+            return claims.get(MicrosoftStsIdToken.PREFERRED_USERNAME);
+        } else if (!StringExtensions.isNullOrBlank(claims.get(MicrosoftStsIdToken.EMAIL))) {
+            return claims.get(MicrosoftStsIdToken.EMAIL);
+        }
+
+        return null;
     }
 
 }
